backend/requirements.txt
bcrypt==3.2.0
black==24.3.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
google-api-core==2.17.1
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
